Implement review comments (Jay)
