+++ components/ai_chat/resources/page/components/drag_overlay/index.tsx
@@ -6,30 +6,78 @@
 import * as React from 'react'
 import { getLocale } from '$web-common/locale'
 import styles from './style.module.scss'
+import { isImageFile } from '../../constants/file_types'
+import * as Mojom from '../../../common/mojom'
 
-import '../../../common/strings'
+const convertFileToUploadedFile = (file: File): Promise<Mojom.UploadedFile> => {
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader()
+    reader.onload = (e) => {
+      const arrayBuffer = e.target?.result as ArrayBuffer
+      if (!arrayBuffer) {
+        reject(new Error('Failed to read file'))
+        return
+      }
+
+      const uint8Array = new Uint8Array(arrayBuffer)
+      const uploadedFile: Mojom.UploadedFile = {
+        filename: file.name,
+        filesize: file.size,
+        data: Array.from(uint8Array),
+        type: Mojom.UploadedFileType.kImage
+      }
+      resolve(uploadedFile)
+    }
+    reader.onerror = () => reject(new Error('Failed to read file'))
+    reader.readAsArrayBuffer(file)
+  })
+}
 
 interface DragOverlayProps {
   isDragActive: boolean
   isDragOver: boolean
-  onDragOver: (e: React.DragEvent) => void
-  onDrop: (e: React.DragEvent) => void
+  clearDragState: () => void
+  processDroppedImages: (files: Mojom.UploadedFile[]) => void
 }
 
 export default function DragOverlay({
   isDragActive,
   isDragOver,
-  onDragOver,
-  onDrop
+  clearDragState,
+  processDroppedImages
 }: DragOverlayProps) {
+  const handleOverlayDrop = async (e: React.DragEvent) => {
+    e.preventDefault()
+    e.stopPropagation()
+    clearDragState()
+
+    const files = Array.from(e.dataTransfer?.files || []).filter(isImageFile)
+
+    if (files.length === 0) {
+      return
+    }
+
+    try {
+      const uploadedFiles = await Promise.all(
+        files.map(file => convertFileToUploadedFile(file))
+      )
+      processDroppedImages(uploadedFiles)
+    } catch (error) {
+      // Silently fail - error will be handled by the upload system
+    }
+  }
+
+  const handleOverlayDragOver = (e: React.DragEvent) => {
+    e.preventDefault()
+  }
   return (
     <>
       {/* Transparent overlay that captures drag events over iframe */}
       {isDragActive && (
         <div
           className={styles.dragDetectionOverlay}
-          onDragOver={onDragOver}
-          onDrop={onDrop}
+          onDragOver={handleOverlayDragOver}
+          onDrop={handleOverlayDrop}
         />
       )}
