+++ components/ai_chat/resources/page/components/drag_overlay/style.module.scss
@@ -49,11 +49,9 @@
 .dragTitle {
   font: var(--leo-font-heading-h4);
   color: var(--leo-color-text-primary);
-  font-weight: 600;
 }
 
 .dragDescription {
   font: var(--leo-font-default-regular);
   color: var(--leo-color-text-secondary);
-  line-height: 1.4;
 }
